llm-service/requirements.txt
# FastAPI and Server
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9

# Google Gemini AI
google-generativeai==0.3.2

# Data Validation
pydantic==2.6.1
pydantic-settings==2.1.0

# YAML Processing
PyYAML==6.0.1

# Security - OWASP Best Practices
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1

# Rate Limiting
slowapi==0.1.9

# HTTP Client
httpx==0.26.0

# Logging and Monitoring
python-json-logger==2.0.7

# Testing
pytest==7.4.4
pytest-asyncio==0.23.4
pytest-cov==4.1.0
httpx==0.26.0

# Code Quality
black==24.1.1
flake8==7.0.0
mypy==1.8.0
